+++ src/Hacks/HitboxTrail.cpp
@@ -1,5 +1,3 @@
-#ifndef GEODE_IS_MACOS
-
 #include <Geode/Geode.hpp>
 #include <Geode/modify/GJBaseGameLayer.hpp>
 #include <Geode/modify/PlayLayer.hpp>
@@ -122,5 +120,3 @@ class $modify (LevelEditorLayer)
         sizes.clear();
     }
 };
-
-#endif