+++ src/Hacks/MenuGameplay.cpp
@@ -1,6 +1,7 @@
-/*#include <Geode/Geode.hpp>
+#include <Geode/Geode.hpp>
 #include <Geode/modify/MenuGameLayer.hpp>
-#include <Geode/modify/CCNode.hpp>
+#include <Geode/modify/MenuLayer.hpp>
+#include <Geode/modify/CCKeyboardDispatcher.hpp>
 #include "../Client/Client.h"
 
 using namespace geode::prelude;
@@ -9,11 +10,14 @@ class MenuGameDelegate : public CCLayer
 {
     public:
         MenuGameLayer* mgl = nullptr;
+        static inline MenuGameDelegate* instance = nullptr;
 
         bool init()
         {
             if (!CCLayer::init())
                 return false;
+
+            instance = this;
 
             this->setKeyboardEnabled(true);
             this->setTouchEnabled(true);
@@ -24,28 +28,29 @@ class MenuGameDelegate : public CCLayer
             return true;
         }
 
-        virtual void keyDown(enumKeyCodes key)
-        {
-            log::info("down");
-        }
-
-		virtual void keyUp(enumKeyCodes key)
-        {
-            log::info("up");
-        }
-
         virtual bool ccTouchBegan(CCTouch *pTouch, CCEvent *pEvent)
         {
-            if (mgl->m_playerObject && !mgl->m_playerObject->m_isSpider)
-                mgl->m_playerObject->pushButton(PlayerButton::Jump);
+            if (Client::GetModuleEnabled("main-menu-gameplay"))
+            {
+                if (mgl->m_playerObject && !mgl->m_playerObject->m_isSpider)
+                    mgl->m_playerObject->pushButton(PlayerButton::Jump);
+            }
 
             return true;
         }
 
-        virtual void ccTouchesEnded(CCSet *pTouches, CCEvent *pEvent)
+        virtual void ccTouchEnded(CCTouch *pTouch, CCEvent *pEvent)
         {
-            if (mgl->m_playerObject && !mgl->m_playerObject->m_isSpider)
-                mgl->m_playerObject->releaseButton(PlayerButton::Jump);
+            if (Client::GetModuleEnabled("main-menu-gameplay"))
+            {
+                if (mgl->m_playerObject && !mgl->m_playerObject->m_isSpider)
+                    mgl->m_playerObject->releaseButton(PlayerButton::Jump);
+            }
+        }
+
+        ~MenuGameDelegate()
+        {
+            instance = nullptr;
         }
 
         CREATE_FUNC(MenuGameDelegate);
@@ -53,11 +58,10 @@ class MenuGameDelegate : public CCLayer
 
 class $modify (MenuGameLayer)
 {   
-    bool tryJump(float p0)
+    void tryJump(float p0)
     {
-        log::info("asdfasdfrewqewrewqfads");
-
-        return MenuGameLayer::tryJump(p0);
+        if (!Client::GetModuleEnabled("main-menu-gameplay"))
+            MenuGameLayer::tryJump(p0);
     }
 
     virtual bool init()
@@ -73,4 +77,33 @@ class $modify (MenuGameLayer)
 
         return true;
     }
-};*/
+};
+
+class $modify (CCKeyboardDispatcher)
+{
+    bool dispatchKeyboardMSG(enumKeyCodes key, bool isKeyDown, bool isKeyRepeat)
+    {
+        if (!isKeyRepeat && MenuGameDelegate::instance)
+        {
+            if (key == as<enumKeyCodes>(38) || key == enumKeyCodes::KEY_W)
+            {
+                if (isKeyDown)
+                    MenuGameDelegate::instance->ccTouchBegan(nullptr, nullptr);
+                else
+                    MenuGameDelegate::instance->ccTouchEnded(nullptr, nullptr);
+            }
+        }
+
+        return CCKeyboardDispatcher::dispatchKeyboardMSG(key, isKeyDown, isKeyRepeat);
+    }
+
+    static void onModify(auto& self) {
+        auto hook = self.getHook("CCKeyboardDispatcher::dispatchKeyboardMSG");
+
+        Loader::get()->queueInMainThread([hook]
+        {
+            auto modu = Client::GetModule("main-menu-gameplay");
+            modu->addHookRaw(hook);
+        });
+    }
+};