+++ src/Hacks/Thicker Hitboxes.cpp
@@ -1,5 +1,3 @@
-#ifndef GEODE_IS_MACOS
-
 #include <Geode/Geode.hpp>
 #include <Geode/modify/GJBaseGameLayer.hpp>
 #include "../Client/Client.h"
@@ -135,5 +133,3 @@ class $modify (GJBaseGameLayer)
             drawForPlayer(m_player2);
     }
 };
-
-#endif